--- RadixWallet/Core/SharedModels/P2P/Dapp/Models/Interaction/DTO/NumberOfAccounts.swift
@@ -1,4 +1,0 @@
-
-// extension P2P.Dapp.Request {
-public typealias NumberOfAccounts = RequestedQuantity
-// }
