--- RadixWallet/Core/SharedModels/P2P/Dapp/Models/Interaction/Requests/AccountsRequestItem.swift
@@ -1,14 +1,0 @@
-// extension P2P.Dapp.Request {
-//	public struct AccountsRequestItem: Sendable, Hashable, Decodable {
-//		public let numberOfAccounts: NumberOfAccounts
-//		public let challenge: DappToWalletInteraction.AuthChallengeNonce?
-//
-//		public init(
-//			numberOfAccounts: NumberOfAccounts,
-//			challenge: DappToWalletInteraction.AuthChallengeNonce?
-//		) {
-//			self.numberOfAccounts = numberOfAccounts
-//			self.challenge = challenge
-//		}
-//	}
-// }
