--- RadixWallet/Core/SharedModels/P2P/Dapp/Models/Interaction/Requests/ResetRequestItem.swift
@@ -1,16 +1,0 @@
-
-// MARK: - P2P.Dapp.Request.ResetRequestItem
-// extension P2P.Dapp.Request {
-//	public struct ResetRequestItem: Sendable, Hashable, Decodable {
-//		public let accounts: Bool
-//		public let personaData: Bool
-//
-//		public init(
-//			accounts: Bool,
-//			personaData: Bool
-//		) {
-//			self.accounts = accounts
-//			self.personaData = personaData
-//		}
-//	}
-// }
