--- RadixWallet/Core/SharedModels/P2P/Dapp/Models/InteractionResponse/DTO/P2P+Dapp+Response+Persona.swift
@@ -1,15 +1,0 @@
-// extension P2P.Dapp.Response {
-//	public struct PersonaResponse: Sendable, Hashable, Encodable {
-//		public let identityAddress: IdentityAddress
-//		public let label: String
-//
-//		public init(identityAddress: IdentityAddress, label: String) {
-//			self.identityAddress = identityAddress
-//			self.label = label
-//		}
-//
-//		public init(persona: Persona) {
-//			self.init(identityAddress: persona.address, label: persona.displayName.rawValue)
-//		}
-//	}
-// }
