--- RadixWallet/Core/SharedModels/P2P/Dapp/Models/InteractionResponse/RequestResponses/PersonaDataRequestResponseItem.swift
@@ -1,21 +1,0 @@
-
-// MARK: - P2P.Dapp.Response.WalletInteractionSuccessResponse.PersonaDataRequestResponseItem
-// WalletToDappInteractionPersonaDataRequestResponseItem
-// extension P2P.Dapp.Response.WalletInteractionSuccessResponse {
-//	public struct PersonaDataRequestResponseItem: Sendable, Hashable, Encodable {
-//		public let name: PersonaDataEntryName?
-//
-//		public let emailAddresses: OrderedSet<PersonaDataEntryEmailAddress>?
-//		public let phoneNumbers: OrderedSet<PersonaDataEntryPhoneNumber>?
-//
-//		public init(
-//			name: PersonaDataEntryName? = nil,
-//			emailAddresses: OrderedSet<PersonaDataEntryEmailAddress>? = nil,
-//			phoneNumbers: OrderedSet<PersonaDataEntryPhoneNumber>? = nil
-//		) {
-//			self.name = name
-//			self.emailAddresses = emailAddresses
-//			self.phoneNumbers = phoneNumbers
-//		}
-//	}
-// }
