+++ RadixWallet/Core/SharedModels/P2P/Dapp/Models/WalletInteraction+PreviewValue.swift
@@ -1,11 +1,4 @@
-// extension DappToWalletInteractionPersonaDataRequestItem {
-//    public static let previewValue = .sample
-// }
-
-// extension P2P.Dapp.Request.SendTransactionItem {
-//	public static let previewValue = try! Self(transactionManifest: .previewValue)
-// }
-//
+// #if DEBUG
 extension WalletInteractionId {
 	public static let previewValue = Self.previewValue0
 	public static let previewValue0: Self = "E621E1F8-C36C-495A-93FC-0C247A3E6E5F"
@@ -15,16 +8,6 @@ extension WalletInteractionId {
 	public static let previewValue4: Self = "A621E1F8-C36C-495A-93FC-0C247A3E6E5F"
 }
 
-//
-// extension DappToWalletInteractionMetadata {
-//	public static let previewValue = try! Self(
-//		version: P2P.Dapp.currentVersion,
-//		networkId: .simulator,
-//		origin: .init(string: "foo.bar"),
-//		dAppDefinitionAddress: .init(validatingAddress: "account_tdx_b_1p8ahenyznrqy2w0tyg00r82rwuxys6z8kmrhh37c7maqpydx7p")
-//	)
-// }
-//
 extension DappToWalletInteraction {
 	public static func previewValueAllRequests() -> Self {
 		.init(
